Add files via upload
Initial stages of the program. This file as of now is able to login to the EAFC web app with the user's email and password, when prompted with an email code the user can enter that into the console, and the program will enter this into the website for them. This can be automated completely but its alot of work for an email right now. Focus for now is to navigate to the transfer market and search it based on user input.